Re factored clan packets
* moved clan packets to theirs own folder
* moved clan handler's to clan region in ClientPacketHandler
