+++ src/Rasa.Game/Packets/Clan/Client/LeaveClanPacket.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Rasa.Packets.MapChannel.Client
+﻿namespace Rasa.Packets.Clan.Client
 {
     using Data;
     using Memory;
