+++ src/Rasa.Game/Packets/Clan/Server/DisplayClanMessagePacket.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Rasa.Packets.MapChannel.Server
+namespace Rasa.Packets.Clan.Server
 {
     using Data;
     using Memory;
